docs: update documentation and codeowners (#7)
* ci: update CODEOWNERS

* ci: Update URL for Contribution Guide Line in greetings.yml

The URL for the Contribution Guide Line in the 'greetings.yml' file within the .github/workflows directory was updated. Previously, no link was provided, only the text. Providing the direct link to the Contribution Guide Line will make it more accessible and user-friendly for the first-time issue creators and PR submitters.

* ci: update CONTRIBUTING.md

* fix: CODEOWNERS

* ci: technical commit

* ci: technical commit