feat: 로또 구매 구현 (#17)
* ci: add ignore file for codecov

* refactor: message 정리

* style: unnecessary trailing comma in test

* feat: 구매할 로또의 개수를 결정하는 LottoCount 구현

* fix: input 검증을 위한 try catch loop 예외 레벨 조정

* style: trailing comma 정의

* style: multiline expression wrapping

* style: new line function signature

* impv: plus, minus 연산 추가

* Add LottoNumbers model and update View and Controller

A feature to validate the number of lotto numbers was created. For this, a LottoNumbers model was added with a from method to transform a list of Integers into LottoNumbers and validate the size. The View is updated to request and display useful lotto-related information. The Controller is also updated to include interactions with the LottoNumbers model. Now, the application is able to manage the user's lotto numbers directly and enforce the correct number of lotto entries.

* style: code formatting

* ci: ignore LottoApp